PoC to demo signal-delivery distribution across threads burning CPU with timer_create.
setitimer wasn't scaling beyond ~20k in our benchmarks, so it was important to have a fallback ready just in case we find HZ unacceptably low.

Now we have it + it mesures thread-wise CPUtime and it delivers signal to the correct thread (best of all worlds), so in profile, we'll see no signal-delivery bias.