Fixed a race that lead to retire-work waiting for processor to stop, which was in-turn waiting for http-raw-writer to free some space in ring so it could write more, but in /profile rpc failure, http-raw-writer does not reading anything! This results in mutual indefinite-wait between producer and retire-flow.
Turns out, this has 3 scenarios:
1. Everything is good and raw-writer transmits data soon after it appears on the ring
2. Backend is slow and recorder does not transmit data fast enough
3. Recorder is partitioned and tx fails

Now scenario 2 and 3 are dealt with by raw-writer failure path, where we set timeouts for scenario 2 and we simply make the ring read-only for both 2 and 3 (so producers, aka processors) come out of write-wait. Scenario 1 is ok anyway wrt this indefinite-wait, because raw-writer will tx and make sure ring is free to take more, hence preventing indefinite-wait.